refactor(server): restore schema-first source surface
Move versioned Hot Updater schema definitions back to a pure schema module so DB renderers and migrators consume the schema instead of owning it.

Constraint: FumaDB dependency must be removed while preserving schema-first/versioned migration behavior and user-facing db commands.

Rejected: keeping version snapshots under db/schema | keeps schema maintenance coupled to implementation code.

Confidence: high

Scope-risk: moderate

Directive: Keep future schema versions under packages/server/src/schema and make provider renderers consume that surface.

Tested: pnpm -w lint:fix; pnpm exec tsc -p packages/server/tsconfig.json --noEmit; pnpm exec vitest run packages/server/src/db/index.spec.ts packages/hot-updater/src/commands/generate.spec.ts packages/hot-updater/src/commands/migrate.spec.ts --reporter=dot; pnpm exec vitest run plugins/plugin-core/src/bundleUnitOfWork.spec.ts plugins/plugin-core/src/createDatabasePlugin.spec.ts packages/server/src/db/pluginCore.spec.ts packages/server/src/handler-standalone.integration.spec.ts --reporter=dot; pnpm --filter @hot-updater/server build

Not-tested: latest-commit standalone hot-updater-agent matrix is pending after this commit.
