+++ packages/server/src/db/hotUpdaterSchema.ts
@@ -1,7 +1,6 @@
-export * from "./schema/definitions";
+export * from "../schema";
 export * from "./schema/mongodb";
 export * from "./schema/registry";
 export * from "./schema/sql";
 export * from "./schema/sqlMigrations";
 export * from "./schema/sqlOperations";
-export * from "./schema/types";
